refactor(conditions): move get_treatment_candidates home to conditions domain
The treat-condition discovery query shared perform_treatment's scene/
engagement/bond gates and queried TreatmentTemplate/ConditionInstance/
PendingAlteration/threads, but lived in the scenes app and reached across
into conditions/services.py's private _scene_participant /
_thread_anchors_to_character helpers. Moved it next to perform_treatment,
where those helpers are locals and TreatmentTargetKind/ConditionInstance/
TreatmentTemplate are already top-level imports; the TARGET_EFFECT_*
discriminators move to conditions/constants.py next to TreatmentTargetKind.

The scenes-app resolver (_resolve_treatment_request) stays put — it is the
consent seam and calls only public perform_treatment + create_interaction.

Task 4 review fix (#1486).
