+++ src/world/scenes/action_services.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from decimal import Decimal
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from django.db import transaction
 from django.utils import timezone
@@ -32,118 +32,10 @@ from world.scenes.types import EnhancedSceneActionResult
 CustomActionResolver = Callable[["SceneActionRequest"], "EnhancedSceneActionResult | None"]
 CUSTOM_ACTION_RESOLVERS: dict[str, CustomActionResolver] = {}
 
-# Discriminator stored on each candidate dict's "target_effect_type" key.
-TARGET_EFFECT_CONDITION = "condition"
-TARGET_EFFECT_ALTERATION = "alteration"
-
 
 def register_custom_action_resolver(action_key: str, resolver: CustomActionResolver) -> None:
     """Register a function that resolves a consent request bypassing the ActionTemplate path."""
     CUSTOM_ACTION_RESOLVERS[action_key] = resolver
-
-
-def get_treatment_candidates(  # noqa: C901
-    helper_sheet: CharacterSheet,
-    target_sheet: CharacterSheet,
-    scene: Scene,
-) -> list[dict[str, Any]]:
-    """Return valid (treatment, target_effect) pairs for helper to attempt on target in scene.
-
-    Each dict contains: treatment, target_effect, target_effect_type ('condition'|'alteration'),
-    bond_thread (if required and valid).
-    """
-    from world.conditions.constants import TreatmentTargetKind  # noqa: PLC0415
-    from world.conditions.models import (  # noqa: PLC0415
-        ConditionInstance,
-        TreatmentTemplate,
-    )
-    from world.conditions.services import (  # noqa: PLC0415
-        _scene_participant,
-        _thread_anchors_to_character,
-    )
-    from world.magic.constants import PendingAlterationStatus  # noqa: PLC0415
-    from world.magic.models import PendingAlteration, Thread  # noqa: PLC0415
-    from world.mechanics.engagement import CharacterEngagement  # noqa: PLC0415
-
-    target_character = target_sheet.character
-    helper_character = helper_sheet.character
-
-    # Engagement gate: neither helper nor target may be engaged.
-    if CharacterEngagement.objects.filter(
-        character__in=[helper_character, target_character]
-    ).exists():
-        return []
-
-    def _scene_ok(treatment: TreatmentTemplate) -> bool:
-        if not treatment.scene_required:
-            return True
-        return (
-            scene.is_active
-            and _scene_participant(scene, helper_sheet)
-            and _scene_participant(scene, target_sheet)
-        )
-
-    candidate_threads = list(Thread.objects.filter(owner=helper_sheet, retired_at__isnull=True))
-
-    def _find_bond_thread() -> Thread | None:
-        for thread in candidate_threads:
-            if _thread_anchors_to_character(thread, target_sheet):
-                return thread
-        return None
-
-    condition_qs = ConditionInstance.objects.filter(
-        target=target_character,
-        resolved_at__isnull=True,
-    ).select_related("condition")
-
-    alteration_qs = PendingAlteration.objects.filter(
-        character=target_sheet,
-        status=PendingAlterationStatus.OPEN,
-    )
-
-    candidates: list[dict[str, Any]] = []
-
-    for treatment in TreatmentTemplate.objects.all():
-        if not _scene_ok(treatment):
-            continue
-
-        if treatment.requires_bond:
-            bond_thread = _find_bond_thread()
-            if bond_thread is None:
-                continue
-        else:
-            bond_thread = None
-
-        if treatment.target_kind == TreatmentTargetKind.PENDING_ALTERATION:
-            candidates.extend(
-                {
-                    "treatment": treatment,
-                    "target_effect": alteration,
-                    "target_effect_type": TARGET_EFFECT_ALTERATION,
-                    "bond_thread": bond_thread,
-                }
-                for alteration in alteration_qs
-            )
-            continue
-
-        for instance in condition_qs:
-            if treatment.target_kind == TreatmentTargetKind.PRIMARY:
-                if instance.condition_id != treatment.target_condition_id:
-                    continue
-            elif treatment.target_kind == TreatmentTargetKind.AFTERMATH:
-                if instance.condition.parent_condition_id != treatment.target_condition_id:
-                    continue
-
-            candidates.append(
-                {
-                    "treatment": treatment,
-                    "target_effect": instance,
-                    "target_effect_type": TARGET_EFFECT_CONDITION,
-                    "bond_thread": bond_thread,
-                }
-            )
-
-    return candidates
 
 
 def _describe_treatment_outcome(
